requirements.txt
openai-whisper
torch
librosa
pydub
noisereduce
soundfile
jiwer
openpyxl
pandas
numpy
python-Levenshtein